+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Please, check out guidelines: https://keepachangelog.com/en/1.0.0/
 * Add support for variant groups https://github.com/xcbuddy/xcbuddy/pull/34 by @pepibumur.
 * Convenience API to `Graph` https://github.com/xcbuddy/xcbuddy/pull/35 by @pepibumur.
 * Target dependencies generation https://github.com/xcbuddy/xcbuddy/pull/36 by @pepibumur.
+* Resources build phase generation https://github.com/xcbuddy/xcbuddy/pull/38 by @pepibumur.
 
 ### Changed